+++ tests/testthat/test_ARCHIGOS.R
@@ -1,4 +1,4 @@
-# Test if the dataset meets the many universe requirements
+# Test if the dataset meets the many packages universe requirements
 
 # Report missing values
 test_that("missing observations are reported correctly", {